refactor: разбиение проекта по папкам
Теперь все внутренние сорсы библиотеки кроме
IController.h в папке src/ (вероятно, интерфейс
тоже нужно переместить в папку src/ и создать
фасад для взаимодействия с библиотекой)

Также теперь отдельно собирается библиотека и
исполняемый файл (target: renderlib и renderlib.x
соответственно)
