+++ src/core/ViewImpl.cpp
@@ -8,8 +8,8 @@
 #include <utility>
 #include <variant>
 
-#include <SDL.h>
-#include <SDL_vulkan.h>
+#include <SDL2/SDL.h>
+#include <SDL2/SDL_vulkan.h>
 
 #include "imgui_impl_sdl2.h"
 #include "imgui_impl_vulkan.h"
