+++ src/core/createInstance.cpp
@@ -1,4 +1,4 @@
-#include "interfaces/IController.h"
+#include "IController.h"
 #include "interfaces/IModel.h"
 #include "core/Controller.h"
 #include "core/Model.h"
