+++ src/graphics/CMakeLists.txt
@@ -1,4 +1,4 @@
-target_sources(SRC
+target_sources(src
         PUBLIC
         vulkan/vk_descriptors.cpp
         vulkan/vk_engine.cpp
@@ -6,12 +6,5 @@ target_sources(SRC
         vulkan/vk_initializers.cpp
         vulkan/vk_loader.cpp
         vulkan/vk_pipelines.cpp
-        vulkan/vk_descriptors.h
-        vulkan/vk_engine.h
-        vulkan/vk_initializers.h
-        vulkan/vk_loader.h
-        vulkan/vk_pipelines.h
-        vulkan/vk_types.h
         Graphics.cpp
-        RenderPathHE.h
 )
