+++ src/graphics/vulkan/vk_images.cpp
@@ -1,5 +1,5 @@
-#include "vk_images.h"
-#include "vk_initializers.h"
+#include "graphics/vulkan/vk_images.h"
+#include "graphics/vulkan/vk_initializers.h"
 
 void vkutil::transition_image(VkCommandBuffer cmd, VkImage image, VkImageLayout currentLayout, VkImageLayout newLayout)
 {
