+++ src/include/core/.gitignore
