+++ src/include/core/Controller.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "interfaces/IController.h"
+#include "IController.h"
 #include "interfaces/IModel.h"
 
 IController::Ptr createController(IModel::Ptr ptr);
