+++ src/include/core/Defines.h
