+++ src/include/core/ViewImpl.h
@@ -2,7 +2,7 @@
 
 #include <memory>
 
-#include "interfaces/IController.h"
+#include "IController.h"
 #include "interfaces/IModel.h"
 #include "interfaces/IView.h"
