+++ src/include/scene/LocalTransformComponent.h
@@ -5,7 +5,7 @@
 #include "glm/vec4.hpp"
 #include "glm/gtx/quaternion.hpp"
 #include "glm/gtx/orthonormalize.hpp"
-#include "../src/core/Logging.h"
+#include "core/Logging.h"
 
 /** @brief A component that stores coordinates in 3D relative to the parent entity.
  *
